itk/agents/python/requirements.txt
a2a-sdk>=0.2.0
uvicorn>=0.30.0
